Kent C. Dodds' Simply React 2018, 2020 Tutorial: State Reducer & Inversion of Control
This React example illustrates a powerful pattern for supporting extensible
components that give behavioral and visual control to the consumer of the component.

This gives you the virtues of scalability and maintainability since clients of the component
are enabled to extend the component without bottlenecking through a maintainer.

It keeps the code-base relatively compact and doesn't burden the maintainer with use-cases
that may be unfamiliar.

I've made some tweaks to the example and captured class-based and hooks-based implementations.