play with concept of layout reducer
I'm trying to extend the notion of state reduction to layout reduction in the service
of 'sub-accordions' that can expand to reveal nested items.  The use case would be
a menu drawer of clickable items that may represent:

  * application routes that affect what is rendered in a content window
  * subdrawers of related items

This commit represents a useful failure.

I'm trying to related the behavior reducers and layout reducers as peers whereas the layout
reducer really should be a downstream /observer/ of the openedItems array managed and mutated
by the behavior reducer.

As is, you get no expand/collapse behavior when clicking on a hand.  The behavior
reducers do fire appropriately, they just has no linkage to the layout reducers and so
the rendered react components remain unchanged.