feat: add notion of a 'selected' accordion item (leaf node)
The useExpandable hook normally just manages an array of indices of currently
expanded items in an Accordion.  For example

    expandedItems = [ 1, 3, 4 ]

With this commit we add a 'focalIndex' to that state.  It represents the last
click-expanded leaf item.  This allows us style it for emphasis (and enable scroll
position tracking later on).

So state now looks like:

    { expandedItems: [ 1, 3, 4], focalIndex: 3 }

This does not break binary compatibility for legacy KCD accordions, but it does
refactor the two underlying hooks (useAccordion and useExpandable) to support the
new state schema.

I'm also cleaning up the usage idiom for combineReducers to have resonable fallback
behavior (to the default expansion reducer) if none of the combined reducers fire.
Generally speaking, the most restrictive reducers should be at the front of the chain.